Crawls multiple sites and links
This crawler looks for all of the links on a website and then looks for all the links on those websites, and so on. It will always search the last website first, then double back. You can specify the max number of links you want to receive from the procedure.